am335x_evm: move this to an capemgr init script

+++ boot/am335x_evm.sh
@@ -24,12 +24,14 @@
 #https://github.com/beagleboard/meta-beagleboard/blob/master/meta-beagleboard-extras/recipes-support/usb-gadget/gadget-init/g-ether-load.sh
 
 #Load new default pinmux:
-if [ -f /lib/firmware/cape-bone-proto-00A0.dtbo ] ; then
-	kernel_version=$(uname -r)
-	if [ ! "x${kernel_version}" = "x3.8.13-bone38" ] ; then
-		capemgr=$(ls /sys/devices/bone_capemgr.*/slots 2> /dev/null)
-		if [ ! "x${capemgr}" = "x" ] ; then
-			echo "cape-bone-proto" > ${capemgr}
+if [ ! -f /etc/default/capemgr ] ; then
+	if [ -f /lib/firmware/cape-bone-proto-00A0.dtbo ] ; then
+		kernel_version=$(uname -r)
+		if [ ! "x${kernel_version}" = "x3.8.13-bone38" ] ; then
+			capemgr=$(ls /sys/devices/bone_capemgr.*/slots 2> /dev/null)
+			if [ ! "x${capemgr}" = "x" ] ; then
+				echo "cape-bone-proto" > ${capemgr}
+			fi
 		fi
 	fi
 fi
